Wda 4.0.2

+++ wserver/src/main/java/com/wanderdrop/wserver/model/Report.java
@@ -15,7 +15,7 @@ public class Report {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name= "report_id")
-    private int reportId;
+    private Long reportId;
 
     @Column(name= "report_heading", nullable = false)
     private String reportHeading;
